refactor(database): criação de todas as tabelas e colunas do banco de dados

+++ TIVIT.CIPA.Api.Domain/Business/ElectionBusiness.cs
@@ -1,5 +1,4 @@
-﻿
-using Azure.Core;
+﻿using Azure.Core;
 using ClosedXML.Excel;
 using DocumentFormat.OpenXml.Spreadsheet;
 using Microsoft.Extensions.Localization;
@@ -37,7 +36,6 @@ namespace TIVIT.CIPA.Api.Domain.Business
             _electionRepository = electionRepository;
             _electionSiteRepository = electionSiteRepository;
             _localizer = localizer;
-
         }
 
         public async Task<Response<ElectionResponse>> GetByIdAsync(int id)
@@ -49,15 +47,21 @@ namespace TIVIT.CIPA.Api.Domain.Business
             if (election == null)
                 return response;
 
-            //essa forma,poderia ser reali 
             var sites = await _electionSiteRepository.GetSitesByElectionIdAsync(id);
 
             response.Data = new ElectionResponse()
             {
                 Id = election.Id,
+                Code = election.Code,
                 Name = election.Name,
+                Description = election.Description,
+                StartDate = election.StartDate,
+                EndDate = election.EndDate,
+                RegistrationStartDate = election.RegistrationStartDate,
+                RegistrationEndDate = election.RegistrationEndDate,
                 ElectionStartDate = election.ElectionStartDate,
                 ElectionEndDate = election.ElectionEndDate,
+                Type = election.Type,
                 IsActive = election.IsActive,
                 Sites = sites.Select(site => new SiteResponse()
                 {
@@ -76,8 +80,8 @@ namespace TIVIT.CIPA.Api.Domain.Business
         {
             var response = new Response<int>();
 
-            var validator = new ElectionValidator(_electionRepository,_localizer);
-             validator.ValidateCreate(createRequest);
+            var validator = new ElectionValidator(_electionRepository, _localizer);
+            validator.ValidateCreate(createRequest);
             if (!validator.IsValid)
             {
                 response.AddMessage(validator.Notifications.Select(x => x.Message));
@@ -86,9 +90,18 @@ namespace TIVIT.CIPA.Api.Domain.Business
 
             var election = new Election()
             {
+                CompanyId = createRequest.CompanyId,
+                Code = createRequest.Code,
                 Name = createRequest.Name,
+                Description = createRequest.Description,
+                StartDate = createRequest.StartDate,
+                EndDate = createRequest.EndDate,
+                RegistrationStartDate = createRequest.RegistrationStartDate,
+                RegistrationEndDate = createRequest.RegistrationEndDate,
                 ElectionStartDate = createRequest.ElectionStartDate,
                 ElectionEndDate = createRequest.ElectionEndDate,
+                Type = createRequest.Type ?? "Regular",
+                InvitationMessage = createRequest.InvitationMessage,
                 IsActive = true,
                 CreateDate = DateTime.Now,
                 CreateUser = _userInfo.Upn
@@ -97,15 +110,17 @@ namespace TIVIT.CIPA.Api.Domain.Business
             await _electionRepository.CreateAsync(election);
 
             #region "Sites da Eleição"
-            var newSites = createRequest.SitesIds.Select(siteId => new ElectionSite()
+            if (createRequest.SitesIds != null && createRequest.SitesIds.Any())
             {
-                SiteId = siteId,
-                ElectionId = election.Id
-            });
+                var newSites = createRequest.SitesIds.Select(siteId => new ElectionSite()
+                {
+                    SiteId = siteId,
+                    ElectionId = election.Id
+                });
 
-            await _electionSiteRepository.AddRangeAsync(newSites);
+                await _electionSiteRepository.AddRangeAsync(newSites);
+            }
             #endregion
-
 
             response.Data = election.Id;
 
@@ -116,8 +131,8 @@ namespace TIVIT.CIPA.Api.Domain.Business
         {
             var response = new Response();
 
-            var validator = new ElectionValidator(_electionRepository,_localizer);
-             validator.ValidateUpdate(id, updateRequest);
+            var validator = new ElectionValidator(_electionRepository, _localizer);
+            validator.ValidateUpdate(id, updateRequest);
             if (!validator.IsValid)
             {
                 response.AddMessage(validator.Notifications.Select(x => x.Message));
@@ -126,29 +141,47 @@ namespace TIVIT.CIPA.Api.Domain.Business
 
             var election = await this._electionRepository.GetByIdAsync(id);
 
+            if (election == null)
+            {
+                response.AddMessage("Eleição não encontrada.");
+                return response;
+            }
+
+            election.Code = updateRequest.Code ?? election.Code;
             election.Name = updateRequest.Name ?? election.Name;
-            election.ElectionStartDate = updateRequest.ElectionStartDate;
-            election.ElectionEndDate = updateRequest.ElectionEndDate;
+            election.Description = updateRequest.Description ?? election.Description;
+            election.StartDate = updateRequest.StartDate ?? election.StartDate;
+            election.EndDate = updateRequest.EndDate ?? election.EndDate;
+            election.RegistrationStartDate = updateRequest.RegistrationStartDate ?? election.RegistrationStartDate;
+            election.RegistrationEndDate = updateRequest.RegistrationEndDate ?? election.RegistrationEndDate;
+            election.ElectionStartDate = updateRequest.ElectionStartDate ?? election.ElectionStartDate;
+            election.ElectionEndDate = updateRequest.ElectionEndDate ?? election.ElectionEndDate;
+            election.Type = updateRequest.Type ?? election.Type;
+            election.InvitationMessage = updateRequest.InvitationMessage ?? election.InvitationMessage;
             election.UpdateDate = DateTime.Now;
             election.UpdateUser = _userInfo.Upn;
 
             await _electionRepository.UpdateAsync(election);
 
             #region "Sites da Eleição"
-
-            var oldSites = await _electionSiteRepository.GetByElectionId(id);
-
-            await _electionSiteRepository.RemoveRangeAsync(oldSites);
-
-            var newSites = updateRequest.SitesIds.Select(siteId => new ElectionSite()
+            if (updateRequest.SitesIds != null && updateRequest.SitesIds.Any())
             {
-                SiteId = siteId,
-                ElectionId = id
-            });
+                var oldSites = await _electionSiteRepository.GetByElectionId(id);
 
-            await _electionSiteRepository.AddRangeAsync(newSites);
+                if (oldSites != null && oldSites.Any())
+                {
+                    await _electionSiteRepository.RemoveRangeAsync(oldSites);
+                }
+
+                var newSites = updateRequest.SitesIds.Select(siteId => new ElectionSite()
+                {
+                    SiteId = siteId,
+                    ElectionId = id
+                });
+
+                await _electionSiteRepository.AddRangeAsync(newSites);
+            }
             #endregion
-
 
             return response;
         }
@@ -167,6 +200,12 @@ namespace TIVIT.CIPA.Api.Domain.Business
 
             var election = await this._electionRepository.GetByIdAsync(id);
 
+            if (election == null)
+            {
+                response.AddMessage("Eleição não encontrada.");
+                return response;
+            }
+
             election.IsActive = isActive;
             election.UpdateDate = DateTime.Now;
             election.UpdateUser = _userInfo.Upn;
@@ -175,6 +214,5 @@ namespace TIVIT.CIPA.Api.Domain.Business
 
             return response;
         }
-
     }
 }
